#112 updated burgers goal oriented demo with both iso- and anisotropic adaptor cases
